// src/app/modules/customers/interfaces/ICustomerDetails.ts
export interface ICustomerDetails {
    customerId: number;
    firstName: string;
    lastName: string;
    email: string;
    mobileNo: string;
    addharNo: string;
    panNo: string;
    dob: Date;
    street: string;
    city: string;
    custState: string;
    zip: number;
    accountNo: number;
    accountTypeId: number;
    balance: number;
    isActive: boolean;
    createdAt: Date;
}